Moved the growing pile of authentication-related stuff into its own package.
