+++ pushy/src/main/java/com/relayrides/pushy/apns/ApnsClientBuilder.java
@@ -31,6 +31,8 @@ import javax.net.ssl.SSLException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.relayrides.pushy.apns.auth.ApnsKeySource;
+import com.relayrides.pushy.apns.auth.ApnsSigningKey;
 import com.relayrides.pushy.apns.proxy.ProxyHandlerFactory;
 
 import io.netty.channel.EventLoopGroup;
