+++ pushy/src/main/java/com/relayrides/pushy/apns/ApnsClientHandler.java
@@ -40,6 +40,9 @@ import org.slf4j.LoggerFactory;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
+import com.relayrides.pushy.apns.auth.ApnsKeyRemovalListener;
+import com.relayrides.pushy.apns.auth.ApnsKeySource;
+import com.relayrides.pushy.apns.auth.ApnsSigningKey;
 
 import io.netty.buffer.ByteBuf;
 import io.netty.channel.ChannelFuture;
