+++ pushy/src/main/java/com/relayrides/pushy/apns/MockApnsServerHandler.java
@@ -19,6 +19,9 @@ import org.slf4j.LoggerFactory;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
+import com.relayrides.pushy.apns.auth.ApnsKeyRemovalListener;
+import com.relayrides.pushy.apns.auth.ApnsKeySource;
+import com.relayrides.pushy.apns.auth.ApnsVerificationKey;
 
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
