+++ pushy/src/main/java/com/relayrides/pushy/apns/auth/ApnsKeyRegistry.java
@@ -1,4 +1,4 @@
-package com.relayrides.pushy.apns;
+package com.relayrides.pushy.apns.auth;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -17,6 +17,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
+
+import com.relayrides.pushy.apns.NoKeyForTopicException;
 
 import io.netty.util.concurrent.Promise;
