+++ pushy/src/main/java/com/relayrides/pushy/apns/auth/ApnsKeyRemovalListener.java
@@ -1,4 +1,4 @@
-package com.relayrides.pushy.apns;
+package com.relayrides.pushy.apns.auth;
 
 public interface ApnsKeyRemovalListener<T extends ApnsKey> {
