+++ pushy/src/main/java/com/relayrides/pushy/apns/auth/ApnsSigningKey.java
@@ -1,4 +1,4 @@
-package com.relayrides.pushy.apns;
+package com.relayrides.pushy.apns.auth;
 
 import java.math.BigInteger;
 import java.security.interfaces.ECPrivateKey;
