+++ pushy/src/main/java/com/relayrides/pushy/apns/auth/ApnsVerificationKeyRegistry.java
@@ -1,4 +1,4 @@
-package com.relayrides.pushy.apns;
+package com.relayrides.pushy.apns.auth;
 
 import java.io.BufferedReader;
 import java.io.IOException;
