+++ pushy/src/test/java/com/relayrides/pushy/apns/auth/KeyPairUtil.java
@@ -1,11 +1,11 @@
-package com.relayrides.pushy.apns;
+package com.relayrides.pushy.apns.auth;
 
 import java.security.KeyPair;
 import java.security.KeyPairGenerator;
 import java.security.NoSuchAlgorithmException;
 import java.security.SecureRandom;
 
-class KeyPairUtil {
+public class KeyPairUtil {
     public static KeyPair generateKeyPair() throws NoSuchAlgorithmException {
         final KeyPairGenerator keyPairGenerator = KeyPairGenerator.getInstance("EC");
         final SecureRandom random = SecureRandom.getInstance("SHA1PRNG");
